refactor(web): extract SenderTaskBadge into its own module
Pulls the purple "From {task}" inter-task provenance badge out of
chat-message.tsx so the queue affordance can reuse the same visual
language for queued inter-task entries. Adds a `size` prop (xs/sm) so
the badge fits both message bubbles and the denser queue rows.
